// client/src/services/openapi/models/ContiGet.ts
export type ContiGet = {
    antipcon: string;
    ancodice: string;
    andescri: string | null;
    ID: string;
    humanReadableID: string;
    andescr2: string | null;
    anindiri: string | null;
    anindir2: string | null;
    anCap: string | null;
    anlocali: string | null;
    anprovin: string | null;
    antelefo: string | null;
    antelfax: string | null;
    annumcel: string | null;
    anperfis: string | null;
    anSesso: string | null;
    andatnas: string | null;
    anlocnas: string | null;
    anpronas: string | null;
    ancodfis: string | null;
    anpariva: string | null;
    anpartsn: string | null;
    ancodba2: string | null;
    an1Mescl: number | null;
    angiosc1: number | null;
    an2Mescl: number | null;
    angiosc2: number | null;
    angiofis: number | null;
    an1Scont: number | null;
    an2Scont: number | null;
    annumlis: string | null;
    antipfat: string | null;
    anbolfat: string | null;
    anprebol: string | null;
    anscorpo: string | null;
    annumcor: string | null;
    anNote: string | null;
    anindweb: string | null;
    anEmail: string | null;
    anEmpec: string | null;
    anritenu: string | null;
    ancognom: string | null;
    anNome: string | null;
    anconsup: string | null;
    antipsot: string | null;
    anccnote: string | null;
    ancctagg: string | null;
    andtinva: string | null;
    andtobso: string | null;
    annotain: number | null;
    antipclf: string | null;
    anflragg: string | null;
    andatavv: string | null;
    anflgavv: string | null;
    antiprif: string | null;
    anconrif: string | null;
    anflesig: string | null;
    anflcodi: string | null;
    afflintr: string | null;
    anconcon: string | null;
    anflfido: string | null;
    anvalfid: number | null;
    anmaxord: number | null;
    anflblve: string | null;
    andatmor: string | null;
    anpagpar: string | null;
    ancoimps: string | null;
    anpeinps: number | null;
    anriinps: number | null;
    ancoinps: number | null;
    ancodatt: number | null;
    ancodtr2: string | null;
    anflgis4: string | null;
    angescon: string | null;
    anflgcon: string | null;
    ancodstu: string | null;
    anibarid: string | null;
    anflaacc: string | null;
    anrisese: string | null;
    ancodass: string | null;
    anclipos: string | null;
    anflesim: string | null;
    ansagint: number | null;
    anflacbd: string | null;
    ancinabi: string | null;
    anIban: string | null;
    anBban: string | null;
    ancodesc: string | null;
    ancodorn: string | null;
    anflimba: string | null;
    anspeinc: number | null;
    ancodirp: string | null;
    ancaurit: string | null;
    ancaspro: number | null;
    anflgcpz: string | null;
    anflsgre: string | null;
    anflinca: string | null;
    ancodsog: string | null;
    ancodcat: string | null;
    ancofisc: string | null;
    anflgest: string | null;
    anflpriv: string | null;
    annumcar: string | null;
    ancodpor: string | null;
    anchksta: string | null;
    anchkmai: string | null;
    anchkpec: string | null;
    anchkfax: string | null;
    anchkwwp: string | null;
    anchkptl: string | null;
    anchkcpz: string | null;
    anflgcau: string | null;
    anconcau: string | null;
    anflivad: string | null;
    anflrite: string | null;
    ancodsal: string | null;
    anErede: string | null;
    aneveecc: string | null;
    anflapca: string | null;
    ancodcuc: string | null;
    ancodreg: string | null;
    antipocl: string | null;
    anchkfir: string | null;
    anidridy: string | null;
    antiidri: number | null;
    anSkype: string | null;
    anivasos: string | null;
    anflblls: string | null;
    anflsoal: string | null;
    anflbodo: string | null;
    anopetre: string | null;
    antippre: string | null;
    anrating: string | null;
    angiorit: number | null;
    anvocfin: string | null;
    anescdof: string | null;
    andespar: string | null;
    anpagfor: string | null;
    ancodsns: string | null;
    andescr3: string | null;
    andescr4: string | null;
    anindir3: string | null;
    anindir4: string | null;
    anlocal2: string | null;
    ancodcom: string | null;
    anscipag: string | null;
    ansprint: string | null;
    ancatpar: string | null;
    anschuma: string | null;
    anforgiu: string | null;
    anflfold: number | null;
    annotcor: string | null;
    antipsog: string | null;
    anflgnet: number | null;
    cpupdtms: string | null;
    anrivbol: string | null;
    anaffida: number | null;
    cpccchk: string | null;
    destinazioni: Array<DesDiveGet>;
};