// client/src/services/openapi/models/DesDiveGet.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type DesDiveGet = {
    ID: string;
    humanReadableID: string;
    ddtipcon: string;
    ddcodice: string;
    ddcoddes: string;
    ddnomdes: string | null;
    ddindiri: string | null;
    ddCap: string | null;
    ddlocali: string | null;
    ddprovin: string | null;
    ddcodnaz: string | null;
    ddperson: string | null;
    ddNote: string | null;
    ddtelefo: string | null;
    ddEmail: string | null;
    ddEmpec: string | null;
    ddtiprif: string | null;
    ddpredef: string | null;
    dddtobso: string | null;
    ddnumfax: string | null;
    ddmccodi: string | null;
    ddmccodt: string | null;
    ddrfdich: string | null;
    ddtelfax: string | null;
    ddtipinf: string | null;
    cpupdtms: string | null;
    cpccchk: string | null;
};